requirements.txt
pydantic>=2.0.0
openai>=1.0.0
pytest>=7.0.0
python-dateutil>=2.8.0
openenv-core[all]>=0.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
